[#180268720] Add three new files (two controllers and one init) and change the app to be accorded.
