+++ controllers/init.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative 'surveyController'
+require_relative 'careerController'
